Feature/websites

+++ src/main/java/pl/edu/pw/pik/pikactivitytrackerserver/model/User.java
@@ -2,6 +2,7 @@ package pl.edu.pw.pik.pikactivitytrackerserver.model;
 
 
 import javax.persistence.*;
+import java.util.UUID;
 
 @Entity
 @Table(name = "users")
@@ -9,11 +10,11 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
+    private int id;
 
     private String login;
     //Contains hash of a password
-    private  String password;
+    private String password;
 
     public User(String login, String password) {
         this.login = login;
